app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.22.0",
    "kagglehub>=0.3.10",
    "pillow>=11.1.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
]
